refactor: Migrate object util to a proper module
