+++ app/script/main/globals.js
@@ -27,7 +27,6 @@ import linkifyGlobal from '../util/linkify.js';
 import htmlGlobal from '../util/linkify-html.js';
 import MessageComparatorGlobal from '../util/MessageComparator.js';
 import NumberUtilGlobal from '../util/NumberUtil.js';
-import ObjectUtilGlobal from '../util/ObjectUtil.js';
 import PeerConnectionUtilGlobal from '../util/PeerConnectionUtil.js';
 import PromiseQueueGlobal from '../util/PromiseQueue.js';
 import helpersGlobal from '../util/scroll-helpers.js';
